DAOS-1021 log: Support batch offering
Support appending multiple entries to the end of the log by adding
two new APIs: raft_append and log_append which accept an array of
entries as arguments. Remove all references to the single entry
versions of the APIs: log_append_entry and raft_append_entry.

Increase capacity of the log buffer exponentially by a factor
of two at a time; enough to ensure that there is enough to
append the given number of entries. This is slightly different
from when appending only a single entry since the buffer size
may have to be increased more than once to be sufficient.

Eliminate the memcpy loop and replace with at the most two calls
for copying existing entries to the newly allocated buffer.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>