DAOS-1310 Build raft using the Makefile target
Currently we build raft by [re]defining it's build recipe in an
SConsscript file. This will leave us continually chasing changes
upstream makes in how it's built.

Instead, we should use the make target supplied by the upstream
Makefile.

While we are at it, we might as well use the Makefile to build the
test_main test tool so that it doesn't have to be done at test run-time.

To continue to support our Variant building with scons, add support to
the raft Makefile for building into a separate named directory.

Add a call to "make clean" when scons -c is run to clean up the build
product.

Add the ability to disable gcov support in the raft Makefile and
disable it during our build from scons.