DAOS-2177 raft: Unit test fails in CI
A raft unit test fails with a segmentation fault because the local
stack variable of 'raft_entry_t' type is not being initialized. It may
happen that the 'type' field ends up with a value from the enumeration
'raft_logtype_e' and gets incorrectly processed by Raft as a change in
the membership, which leads to the segfault.

Signed-off-by: Omkar Kulkarni <omkar.kulkarni@intel.com>