DAOS-2054 server: Always go through elections
An optimization in raft_periodic lets a node who is the only voting node
in its local configuration become a leader without any election. No
persistent term or vote update is involved.

This behavior violates some Raft rules. Consider a two-node
configuration, C = {S1, S2}, where S1 is a follower in term 1 ("F1") and
S2 is a leader in term 1 ("L1"):

  S1 (F1): [C]
  S2 (L1): [C]

Both S1 and S2 have voted for S2 in term 1. Then, S2 is removed,
resulting in a new configuration, D = {S1}. Before S2 receives a reply
from S1 indicating that S1 has D:

  S1 (F1): [C, D]
  S2 (L1): [C, D]

S1 may become a leader in term 1 through the optimization, because it is
the only voting node in D, it doesn't increment the term, and it doesn't
go through any election. Before S2 commits D and steps down, the two
leaders in term 1 may even exist simultaneously.

Another concern is that if S1 above restarts later, it will become the
leader of term 1 again. One may argue that this should be a different
leadership in a different term, for certain volatile states of the user
or raft itself (e.g., nextIndex[] and matchIndex[]) are reset.

After all, the performance benefit from the optimization does not seem
to justify the correctness concerns (or the logical complexity, at
least, if one regards two-node configurations as unimportant). It is
with these considerations, the patch removes the optimization.

Signed-off-by: Li Wei <wei.g.li@intel.com>