DAOS-9326 server: Fix membership changes
(The core change here may be of no interest to the upstream.)

Unfortunately, daos ends up needing to support re-adding a node with its
previous ID, which does not work properly with our algorithm and is not
supported by the upstream algorithm too. While daos will avoid reusing
node IDs eventually, it needs to provide compatiblity for existing
persistent raft state. This patch therefore reverts to the upstream
algorithm but drops the troublemaking logic that tracks the last applied
membership, which is not used by daos. The resulting algorithm is based
on the following principles:

  - raft only tracks the active membership; the user shall track the
    last applied membership for snapshotting purposes. (Tracking both
    using the same set of raft_node_t objects in raft proves to be
    complex. The user does get a burden, but a very clear one.)
      * The INACTIVE flag allows a node to be absent in the active
	membership but still present in the last applied one.
      * The ADDITION_COMMITTED and VOTING_COMMITTED flags track the last
	applied membership. The former corresponds to the INACTIVE flag
        for the active membership; the latter to the VOTING flag.

  - Each membership change entry is a clear transition from a fixed
    pre-state to a fixed post-state. For instance, ADD_NODE changes a
    node from absent to voting. This enables clear undo logic in
    raft_pop_log.
      * PROMOTE_NODE and REMOVE_NON_VOTING_NODE are introduced to enable
	upgrading from existing persistent raft state that doesn't use
        non-voting nodes.

This patch also makes the following minor changes:

  - Fix an upstream bug that does not consider REMOVE_NODE as a voting
    configuration change.

  - Add a check to raft_recv_entry for invalid membership change
    entries.
      * Removing a leader is currently not safe and now prohibited.
      * Enforce the pre-state of each membership change.

  - Remove a redundant setting of voting_cfg_change_log_idx in
    raft_clear.

  - Change raft_add_node to disallow "promotion" semantics, which shall
    be unnecessary to the user.

To control the size of the patch, the revision of node pointer usage and
some other related fixes are not included (but may be required).

Signed-off-by: Li Wei <wei.g.li@intel.com>