DAOS-10249 server: Fix RV assertion failures
A failure of the following assertion in raft_recv_requestvote was
observed:

  assert(!raft_is_leader(me_) &&
         (vr->prevote || !raft_is_candidate(me_)));

At the time, this server was very unlikely a leader. So it must be a
candidate granting a non-prevote RV request. Since a candidate votes for
itself only when becoming a _prevoted_ candidate, the server must be a
_prevoting_ candidate. This patch fixes the assertion to allow such a
valid case.

Related, raft_become_prevoted_candidate clears me->prevote too early. If
the raft_set_current_term call or the raft_vote_for_nodeid call returns
an error, the server is left as a prevoted candidate, but hasn't voted
for itself. If a non-prevote RV request comes in, the server may grant
the vote and fail the assertion above. This patch changes
raft_become_prevoted_candidate to exit the prevoting phase after the
self vote succeeds.

Signed-off-by: Li Wei <wei.g.li@intel.com>