DAOS-10139 include: Allow 64-bit node IDs (#82)
Change raft_node_id_t from int to long so that it'll be easier for users
to generate unique node IDs. (Using conditional compilation turns out to
be tricky, because users' build system must define consistent macro
values throughout. I believe that we must introduce a configuration
phase for this library to support that well.)

Signed-off-by: Li Wei <liwei@hpe.com>